requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
backports.functools-lru-cache==2.0.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cn2an==0.5.22
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
Cython==0.29.21
dataclasses-json==0.6.4
decorator==4.4.2
demucs==4.0.1
distro==1.9.0
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dora_search==0.1.12
einops==0.7.0
eng-to-ipa==0.0.2
exceptiongroup==1.2.0
filelock==3.13.4
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
indic-transliteration==2.3.37
inflect==7.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.3
joblib==1.4.0
jsonpatch==1.33
jsonpointer==2.4
julius==0.2.7
kiwisolver==1.4.5
ko-pron==1.3
lameenc==1.7.0
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.45
langchain-text-splitters==0.0.1
langsmith==0.1.49
librosa==0.9.2
llvmlite==0.42.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.4
mdurl==0.1.2
more-itertools==10.2.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
num-thai==0.0.5
numba==0.59.1
numpy==1.26.4
omegaconf==2.3.0
openai==1.14.2
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
OpenCC==1.1.1
openunmix==1.3.0
orjson==3.10.1
packaging==23.2
pillow==10.3.0
platformdirs==4.2.0
pooch==1.8.1
proces==0.1.7
proglog==0.1.10
protobuf==5.26.1
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
PyJWT==2.8.0
pyopenjtalk-prebuilt==0.3.0
pyparsing==3.1.2
pypinyin==0.51.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
resampy==0.4.3
retrying==1.3.4
rich==13.7.1
roman==4.1
scikit-learn==1.0.2
scipy==1.13.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
SQLAlchemy==2.0.29
sqlparse==0.4.4
submitit==1.5.1
sympy==1.12
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.4.0
tiktoken==0.6.0
toml==0.10.2
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
tqdm==4.66.2
treetable==0.2.5
typeguard==4.2.1
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.1
Werkzeug==3.0.2
yarl==1.9.4